microservices/article/requirements.txt
Flask==3.0.*
Flask-SQLAlchemy==3.1.*
flask_marshmallow==1.3.0
marshmallow==3.21.*
marshmallow-sqlalchemy==0.29.*
psycopg2-binary==2.9.*
python-slugify==8.*
python-dotenv==1.*